1/requirements.txt
django-haystack==2.4.1
Whoosh==2.7.0
